int casting hope that fixes it

+++ src/api/bottler.py
@@ -57,7 +57,7 @@ def get_bottle_plan():
                     return [
                         {
                             "potion_type": [100, 0, 0, 0],
-                            "quantity": (num_red_ml / 100),
+                            "quantity": int(num_red_ml / 100),
                         }
                 ]
         return []
